Make JavaCompatiblePartitioner new default
The JavaCompatiblePartitioner is renamed DefaultPartitioner. The old
DefaultPartitioner becomes LegacyPartitioner.
