+++ src/producer/partitioners/default/index.js
@@ -1,4 +1,4 @@
 const murmur2 = require('./murmur2')
-const createDefaultPartitioner = require('./partitioner')
+const createDefaultPartitioner = require('../legacy/partitioner')
 
 module.exports = createDefaultPartitioner(murmur2)
