+++ src/producer/partitioners/default/murmur2.spec.js
@@ -8,29 +8,31 @@ describe('Producer > Partitioner > Default > murmur2', () => {
   })
 
   test('it handles numeric input', () => {
-    expect(murmur2(0)).toEqual(272173970)
+    expect(murmur2(0)).toEqual(971027396)
   })
 
   test('it handles buffer input', () => {
-    expect(murmur2(Buffer.from('1'))).toEqual(1311020360)
+    expect(murmur2(Buffer.from('1'))).toEqual(-1993445489)
   })
 })
 
+// Generated with src/producer/partitioners/defaultJava/Test.java
 const testData = {
-  '0': 272173970,
-  '1': 1311020360,
-  '128': 2053105854,
-  '2187': -2081355488,
-  '16384': 204404061,
-  '78125': -677491393,
-  '279936': -622460209,
-  '823543': 651276451,
-  '2097152': 944683677,
-  '4782969': -892695770,
-  '10000000': -1778616326,
-  '19487171': -518311627,
-  '35831808': 556972389,
-  '62748517': -233806557,
-  '105413504': -109398538,
-  '170859375': 102939717,
+  '0': 971027396,
+  '1': -1993445489,
+  '128': -326012175,
+  '2187': -1508407203,
+  '16384': -325739742,
+  '78125': -1654490814,
+  '279936': 1462227128,
+  '823543': -2014198330,
+  '2097152': 607668903,
+  '4782969': -1182699775,
+  '10000000': -1830336757,
+  '19487171': -1603849305,
+  '35831808': -857013643,
+  '62748517': -1167431028,
+  '105413504': -381294639,
+  '170859375': -1658323481,
+  '100:48069': 1009543857,
 }
