--- src/producer/partitioners/defaultJava/index.js
@@ -1,4 +1,0 @@
-const murmur2 = require('./murmur2')
-const createDefaultPartitioner = require('../default/partitioner')
-
-module.exports = createDefaultPartitioner(murmur2)
