+++ src/producer/partitioners/legacy/index.js
@@ -1,0 +1,4 @@
+const murmur2 = require('./murmur2')
+const createLegacyPartitioner = require('./partitioner')
+
+module.exports = createLegacyPartitioner(murmur2)
