+++ src/producer/partitioners/legacy/partitioner.js
